Добавлен реакт.контекст, произведен рефакторинг попапов, добавлена функция добавления карточки

+++ src/utils/Api.js
@@ -47,20 +47,20 @@ class Api {
     });
   }
 
-  sendUserInfo(profileData) {
+  setUserInfo(data) {
     return fetch(`${this._url}users/me`, {
       headers: this._headers,
       method: 'PATCH',
       body: JSON.stringify({
-        name: profileData.userName,
-        about: profileData.userDescription,
+        name: data.name,
+        about: data.about,
       }),
     }).then((res) => {
       return this._checkResponse(res);
     });
   }
 
-  sendAvatar(avatarLink) {
+  setAvatar(avatarLink) {
     return fetch(`${this._url}users/me/avatar`, {
       headers: this._headers,
       method: 'PATCH',
@@ -70,22 +70,22 @@ class Api {
     });
   }
 
-  putLike(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: 'PUT',
-    }).then((res) => {
-      return this._checkResponse(res);
-    });
-  }
-
-  deleteLike(cardId) {
-    return fetch(`${this._url}cards/${cardId}/likes`, {
-      headers: this._headers,
-      method: 'DELETE',
-    }).then((res) => {
-      return this._checkResponse(res);
-    });
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: 'PUT',
+      }).then((res) => {
+        return this._checkResponse(res);
+      });
+    } else {
+      return fetch(`${this._url}cards/${cardId}/likes`, {
+        headers: this._headers,
+        method: 'DELETE',
+      }).then((res) => {
+        return this._checkResponse(res);
+      });
+    }
   }
 }
